changed total cost column

+++ db/migrate/20250117164226_create_orders.rb
@@ -6,7 +6,7 @@ class CreateOrders < ActiveRecord::Migration[7.2]
       t.string :status
       t.text :event_details
       t.integer :group_size
-      t.integer :total_cost
+      t.decimal :total_cost
 
       t.timestamps
     end
